Move Application to another project
