+++ TagCloud/Result.cs
@@ -23,7 +23,7 @@ namespace TagCloud
         }
 
         public string Error { get; }
-        internal T Value { get; }
+        public T Value { get; }
 
         public T GetValueOrThrow()
         {
