+++ TagCloud/TagCloud.csproj
@@ -2,29 +2,31 @@
 
     <PropertyGroup>
         <LangVersion>latest</LangVersion>
-        <OutputType>Exe</OutputType>
         <TargetFramework>net6.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Autofac" Version="6.3.0" />
-      <PackageReference Include="CommandLineParser" Version="2.9.0-preview1" />
-      <PackageReference Include="DeepMorphy" Version="2.0.3" />
-      <PackageReference Include="FluentAssertions" Version="6.2.0" />
-      <PackageReference Include="NHunspell.Patched" Version="1.2.5554" />
-      <PackageReference Include="NUnit" Version="3.13.2" />
-      <PackageReference Include="System.Drawing.Common" Version="4.7.0" />
+        <PackageReference Include="Autofac" Version="6.3.0"/>
+        <PackageReference Include="CommandLineParser" Version="2.9.0-preview1"/>
+        <PackageReference Include="DeepMorphy" Version="2.0.3"/>
+        <PackageReference Include="FluentAssertions" Version="6.2.0"/>
+        <PackageReference Include="NHunspell" Version="1.2.5554.16953"/>
+        <PackageReference Include="NUnit" Version="3.13.2"/>
+        <PackageReference Include="System.Drawing.Common" Version="4.7.0"/>
     </ItemGroup>
     <ItemGroup>
-        <Content Include="ru_RU.aff">
+        <EmbeddedResource Include="Resources\Hunspellx64.dll">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="ru_RU.dic">
+        </EmbeddedResource>
+        <EmbeddedResource Include="Resources\Hunspellx86.dll">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-    </ItemGroup>
-    <ItemGroup>
-      <Compile Remove="TextHandlers\Converters\PunctuationMarksRemover.cs" />
+        </EmbeddedResource>
+        <EmbeddedResource Include="Resources\ru_RU.aff">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </EmbeddedResource>
+        <EmbeddedResource Include="Resources\ru_RU.dic">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </EmbeddedResource>
     </ItemGroup>
 
 </Project>
