+++ TagCloud/TextHandlers/Converters/Stemer.cs
@@ -5,7 +5,7 @@ namespace TagCloud.TextHandlers.Converters
 {
     public class Stemer : IConverter
     {
-        private readonly Hunspell hunspell = new("../../../ru_ru.aff", "../../../ru_ru.dic");
+        private readonly Hunspell hunspell = new("ru_ru.aff", "ru_ru.dic");
 
         public Result<string> Convert(string original)
         {
